[futures::stream] concurrent_map impl
