+++ Cargo.toml
@@ -507,10 +507,10 @@ event-listener = "2.5.3"
 fail = "0.5.0"
 field_count = "0.1.1"
 flate2 = "1.0.24"
-futures = "0.3.24"
-futures-channel = "0.3.24"
-futures-core = "0.3.25"
-futures-util = "0.3.21"
+futures = "0.3.29"
+futures-channel = "0.3.29"
+futures-core = "0.3.29"
+futures-util = "0.3.29"
 gcp-bigquery-client = "0.13.0"
 get_if_addrs = "0.5.3"
 git2 = "0.16.1"
@@ -612,6 +612,7 @@ rocksdb = { version = "0.21.0", features = ["lz4"] }
 rstack-self =  { version = "0.3.0", features = ["dw"], default_features = false }
 rstest = "0.15.0"
 rusty-fork = "0.3.0"
+rustversion = "1.0.14"
 scopeguard = "1.2.0"
 sha-1 = "0.10.0"
 sha2 = "0.9.3"
