+++ crates/bounded-executor/Cargo.toml
@@ -14,6 +14,7 @@ rust-version = { workspace = true }
 
 [dependencies]
 futures = { workspace = true }
+rustversion = { workspace = true }
 tokio = { workspace = true }
 
 [dev-dependencies]
